backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
python-multipart
httpx
pytest